requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
coloredlogs==15.0.1
ddddocr==1.4.8
dnspython==2.4.2
exceptiongroup==1.1.3
flatbuffers==23.5.26
h11==0.14.0
humanfriendly==10.0
idna==3.4
mpmath==1.3.0
numpy==1.26.0
onnxruntime==1.16.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
outcome==1.2.0
packaging==23.2
pandas==2.1.1
Pillow==9.5.0
protobuf==4.24.3
pycparser==2.21
pymongo==4.5.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.12
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==1.26.17
webdriver-manager==4.0.1
wsproto==1.2.0
